+++ veryfi/errors.py
@@ -42,6 +42,10 @@ class BadRequest(VeryfiClientError):
     pass
 
 
+class ResourceNotFound(VeryfiClientError):
+    pass
+
+
 class UnexpectedHTTPMethod(VeryfiClientError):
     pass
 
@@ -56,6 +60,7 @@ class InternalError(VeryfiClientError):
 
 _error_map = {
     400: BadRequest,
+    404: ResourceNotFound,
     401: UnauthorizedAccessToken,
     405: UnexpectedHTTPMethod,
     409: AccessLimitReached,